+++ docarray/__init__.py
@@ -1,7 +1,7 @@
 __version__ = '0.1.0'
 
 from docarray.array.array import DocumentArray
-from docarray.document_base.document import BaseDocument
+from docarray.base_document.document import BaseDocument
 
 __all__ = [
     'BaseDocument',
