+++ docarray/array/array_stacked.py
@@ -15,7 +15,7 @@ from typing import (
 
 from docarray.array.abstract_array import AnyDocumentArray
 from docarray.array.array import DocumentArray
-from docarray.document_base import AnyDocument, BaseDocument
+from docarray.base_document import AnyDocument, BaseDocument
 from docarray.typing import NdArray
 from docarray.typing.tensor.abstract_tensor import AbstractTensor
 from docarray.utils._typing import is_tensor_union
