+++ docarray/base_document/__init__.py
@@ -1,0 +1,5 @@
+from docarray.base_document.any_document import AnyDocument
+from docarray.base_document.base_node import BaseNode
+from docarray.base_document.document import BaseDocument
+
+__all__ = ['AnyDocument', 'BaseDocument', 'BaseNode']
