+++ docarray/base_document/base_node.py
