+++ docarray/base_document/io/json.py
