+++ docarray/base_document/mixins/__init__.py
@@ -1,0 +1,3 @@
+from docarray.base_document.mixins.proto import ProtoMixin
+
+__all__ = ['ProtoMixin']
