+++ tests/integrations/document/test_to_json.py
@@ -1,7 +1,7 @@
 import numpy as np
 import torch
 
-from docarray.document_base import BaseDocument
+from docarray.base_document import BaseDocument
 from docarray.typing import AnyUrl, NdArray, TorchTensor
