+++ tests/units/typing/tensor/test_tensor.py
@@ -3,7 +3,7 @@ import orjson
 import pytest
 from pydantic.tools import parse_obj_as, schema_json_of
 
-from docarray.document_base.io.json import orjson_dumps
+from docarray.base_document.io.json import orjson_dumps
 from docarray.typing import NdArray
 from docarray.typing.tensor import NdArrayEmbedding
