+++ tests/units/typing/tensor/test_torch_tensor.py
@@ -2,7 +2,7 @@ import pytest
 import torch
 from pydantic.tools import parse_obj_as, schema_json_of
 
-from docarray.document_base.io.json import orjson_dumps
+from docarray.base_document.io.json import orjson_dumps
 from docarray.typing import TorchEmbedding, TorchTensor
