+++ tests/units/typing/url/test_text_url.py
@@ -4,7 +4,7 @@ import urllib
 import pytest
 from pydantic import parse_obj_as, schema_json_of
 
-from docarray.document_base.io.json import orjson_dumps
+from docarray.base_document.io.json import orjson_dumps
 from docarray.typing import TextUrl
 
 REMOTE_TXT = 'https://de.wikipedia.org/wiki/Brixen'
